services/billing/requirements.txt
celery==5.2.6
Django==3.2.9
django-celery-beat==2.2.1
django-celery-results==2.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
django-cors-headers==3.12.0
drf-spectacular==0.22.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
pytest-django==4.5.2
redis==4.3.0
jinja2==3.1.2
requests==2.27.1
pydantic==1.9.0
orjson==3.6.8
aiohttp==3.8.1
flower==1.0.0
python-dotenv==0.20.0
